+++ src/events.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { createEmitters, createTypes, createUnion, eol } from './common'
+import { createEmitters, createEmittersV2, createTypes, createUnion, eol } from './common'
 import { Definitions, TextPos } from './definitionParser'
 
 export const buildEvents = async (editor: vscode.TextEditor, pos: TextPos, events: Definitions) => {
@@ -8,6 +8,22 @@ export const buildEvents = async (editor: vscode.TextEditor, pos: TextPos, event
     edit.insert(
       pos,
       createTypes('Event', events) + createUnion('Event', events) + n + createEmitters(events),
+    )
+  })
+}
+export const buildEventsV2 = async (
+  editor: vscode.TextEditor,
+  pos: TextPos,
+  events: Definitions,
+) => {
+  await editor.edit((edit) => {
+    const n = eol(editor.document)
+    edit.insert(
+      pos,
+      createTypes('Event', events) +
+        createUnion('Event', events) +
+        n +
+        createEmittersV2(editor.document, events),
     )
   })
 }